pyproject.toml
[project]
name = "agent-key"
version = "0.1.0"
description = "API key management service for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.1",
    "fastapi>=0.117.1",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.36.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--no-header -rN"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
